- moved ReservationOperationsFactory to the HotelManager constructor argument - changed ReservationService folder name to no be the same as ReservationService class
