+++ HotelBookingTests/HotelExampleBookRoomCantFailTest.cs
@@ -2,7 +2,7 @@
 using HotelBooking;
 using HotelBooking.HotelExamples;
 using HotelBooking.ReservationOperationsProvider;
-using HotelBooking.ReservationService;
+using HotelBooking.ReservationServices;
 using NUnit.Framework;
 using Rhino.Mocks;
