+++ HotelBookingTests/HotelExampleTwoStepPaymentTest.cs
@@ -4,7 +4,7 @@ using HotelBooking;
 using HotelBooking.HotelExamples;
 using HotelBooking.PaymentExternalService;
 using HotelBooking.ReservationOperationsProvider;
-using HotelBooking.ReservationService;
+using HotelBooking.ReservationServices;
 using NUnit.Framework;
 using Rhino.Mocks;
