+++ HotelBookingTests/ReservationServiceOneStepPaymentTest.cs
@@ -5,7 +5,7 @@ using System.Text.RegularExpressions;
 using HotelBooking;
 using HotelBooking.HotelExamples;
 using HotelBooking.ReservationOperationsProvider;
-using HotelBooking.ReservationService;
+using HotelBooking.ReservationServices;
 using NUnit.Framework;
 using Rhino.Mocks;
